refactor(backend): split lib.rs into per-domain modules (3659 → 455 LoC)
v1.1 left src-tauri/src/lib.rs as a ~3,700-line god-object holding
every Tauri command alongside its types, path helpers, persistence,
and validation code. Splitting it unblocks the rest of bucket B
(targeted tests, duplicate-helper consolidation, future contributors).

New layout:

- types.rs              — all serde structs (NoteFile, FolderInfo,
                          TrashedNote, Template, ContentMatch, etc.)
- paths.rs              — get_*_dir helpers, ensure_*_dir, file_modified_unix
- validation.rs         — is_safe_filename, validate_path_within_base,
                          validate_user_export_path
- persist.rs            — config / trash-metadata / note-metadata IO,
                          generate_unique_{filename,folder_name}
- wiki.rs               — parse_wiki_links, note_name_to_filename,
                          note_exists, get_link_context
- templates_data.rs     — default templates + variable substitution
- commands/notes.rs     — note CRUD + list + move + export_single + rename
- commands/folders.rs   — folder CRUD + move + rename + recursive scan
- commands/trash.rs     — trash / restore / cleanup / empty / preview
- commands/templates.rs — template CRUD + apply / create_from_template
- commands/locking.rs   — lock / unlock / permanently_unlock / is_locked
- commands/search.rs    — search_notes_content + snippet helpers
- commands/backlinks.rs — scan_note_links + get_backlinks + create_from_link
- commands/export_import.rs — ZIP + encrypted backup + settings JSON
- commands/misc.rs      — directory mgmt, colors, binary file write, images

lib.rs now contains only module declarations, the macOS-gated calendar
commands, `pub fn run()` with the full invoke_handler!, and the unit
test module.

Pure mechanical move — no function signatures, behavior, or tests changed.
Visibility widened to pub(crate) only where required for cross-module access.

Verification:
- cargo clippy --lib --tests -- -D warnings → clean
- cargo test --lib → 24 passed, 0 failed
- npm run build → passes
