+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,11 @@
+//! Tauri command implementations, grouped by domain.
+
+pub mod backlinks;
+pub mod export_import;
+pub mod folders;
+pub mod locking;
+pub mod misc;
+pub mod notes;
+pub mod search;
+pub mod templates;
+pub mod trash;
