+++ src-tauri/src/wiki.rs
@@ -1,0 +1,109 @@
+//! Wiki link parsing and resolution helpers.
+
+use lazy_static::lazy_static;
+use regex::Regex;
+
+use crate::paths::get_notes_dir;
+
+// Wiki Link Regex
+lazy_static! {
+    // Matches [[Note Name]] or [[Display|note-name]]
+    static ref WIKI_LINK_REGEX: Regex = Regex::new(r"\[\[([^\]|]+)(?:\|([^\]]+))?\]\]").unwrap();
+}
+
+pub(crate) fn parse_wiki_links(content: &str) -> Vec<String> {
+    let mut links = Vec::new();
+
+    for cap in WIKI_LINK_REGEX.captures_iter(content) {
+        // Get the target (second capture group if exists, otherwise first)
+        let target = cap
+            .get(2)
+            .or_else(|| cap.get(1))
+            .map(|m| m.as_str().to_string())
+            .unwrap_or_default();
+
+        if !target.is_empty() {
+            links.push(target);
+        }
+    }
+
+    links
+}
+
+pub(crate) fn note_name_to_filename(note_name: &str) -> String {
+    // Convert "Meeting Notes" -> "meeting-notes.md"
+    let slug = note_name
+        .to_lowercase()
+        .trim()
+        .replace(' ', "-")
+        .replace(|c: char| !c.is_alphanumeric() && c != '-', "");
+
+    format!("{}.md", slug)
+}
+
+pub(crate) fn note_exists(note_name: &str) -> Result<(bool, String), String> {
+    let notes_dir = get_notes_dir();
+
+    // Try as standalone note first
+    let filename = note_name_to_filename(note_name);
+    let standalone_path = notes_dir.join("notes").join(&filename);
+
+    if standalone_path.exists() {
+        return Ok((true, filename));
+    }
+
+    // Try as daily note (YYYY-MM-DD format)
+    let daily_filename = if note_name.ends_with(".md") {
+        note_name.to_string()
+    } else {
+        format!("{}.md", note_name)
+    };
+    let daily_path = notes_dir.join("daily").join(&daily_filename);
+    if daily_path.exists() {
+        return Ok((true, daily_filename));
+    }
+
+    Ok((false, filename))
+}
+
+pub(crate) fn get_link_context(content: &str, link_text: &str) -> String {
+    // Try both with and without pipe syntax
+    let search_patterns = vec![
+        format!("[[{}]]", link_text),
+        format!("[[{}|", link_text),
+    ];
+
+    for search in search_patterns {
+        if let Some(pos) = content.find(&search) {
+            let start = pos.saturating_sub(50);
+            let end = (pos + search.len() + 50).min(content.len());
+
+            // Find the actual end of the link
+            let actual_end = if search.ends_with('|') {
+                // Find the closing ]]
+                content[pos..]
+                    .find("]]")
+                    .map(|p| (pos + p + 2 + 50).min(content.len()))
+                    .unwrap_or(end)
+            } else {
+                end
+            };
+
+            let context = &content[start..actual_end];
+
+            // Add ellipsis if truncated
+            let mut result = String::new();
+            if start > 0 {
+                result.push_str("...");
+            }
+            result.push_str(context);
+            if actual_end < content.len() {
+                result.push_str("...");
+            }
+
+            return result;
+        }
+    }
+
+    String::new()
+}
